+++ examples/fiveMinuteGuide.js
@@ -18,32 +18,32 @@ const bob = `bob-${uuid}`
 
 async function main() {
   const key = await ledger.keys.create()
-  await ledger.assets.create({ alias: usd, keyIds: [key.id] })
+  await ledger.flavors.create({ id: usd, keyIds: [key.id] })
   await ledger.accounts.create({ id: alice, keyIds: [key.id] })
   await ledger.accounts.create({ id: bob, keyIds: [key.id] })
 
   await ledger.transactions.transact(builder => {
     builder.issue({
-      asset_alias: usd,
+      flavorId: usd,
       amount: 100,
-      destination_account_id: alice,
+      destinationAccountId: alice,
     })
   })
 
   await ledger.transactions.transact(builder => {
     builder.transfer({
-      asset_alias: usd,
+      flavorId: usd,
       amount: 50,
-      source_account_id: alice,
-      destination_account_id: bob,
+      sourceAccountId: alice,
+      destinationAccountId: bob,
     })
   })
 
   await ledger.transactions.transact(builder => {
     builder.retire({
-      asset_alias: usd,
+      flavorId: usd,
       amount: 20,
-      source_account_id: bob,
+      sourceAccountId: bob,
     })
   })
 }