+++ examples/fiveMinuteGuide.ts
@@ -21,13 +21,13 @@ const bob = `bob-${uuid}`
 
 async function main() {
   const key = await ledger.keys.create()
-  await ledger.assets.create({ alias: usd, keyIds: [key.id] })
+  await ledger.flavors.create({ id: usd, keyIds: [key.id] })
   await ledger.accounts.create({ id: alice, keyIds: [key.id] })
   await ledger.accounts.create({ id: bob, keyIds: [key.id] })
 
   await ledger.transactions.transact((builder: TransactionBuilder) => {
     builder.issue({
-      assetAlias: usd,
+      flavorId: usd,
       amount: 100,
       destinationAccountId: alice,
     })
@@ -35,7 +35,7 @@ async function main() {
 
   await ledger.transactions.transact((builder: TransactionBuilder) => {
     builder.transfer({
-      assetAlias: usd,
+      flavorId: usd,
       amount: 50,
       sourceAccountId: alice,
       destinationAccountId: bob,
@@ -44,7 +44,7 @@ async function main() {
 
   await ledger.transactions.transact((builder: TransactionBuilder) => {
     builder.retire({
-      assetAlias: usd,
+      flavorId: usd,
       amount: 20,
       sourceAccountId: bob,
     })