+++ src/api/keys.ts
@@ -12,10 +12,6 @@ import { Consumer, ObjectCallback, PageCallback, sharedAPI } from '../shared'
  * @typedef {Object} Key
  * @global
  *
- * @property {String} alias
- * **Deprecated. Use id instead.**
- * User specified, unique identifier of the key.
- *
  * @property {String} id
  * Unique identifier of the key.
  */
@@ -35,12 +31,10 @@ export const keysAPI = (client: Client) => {
      * @param {Object} params - Parameters for key creation.
      * @param {String} params.id - Unique identifier. Will be auto-generated if
      *   not provided.
-     * @param {String} params.alias - **Deprecated. Use id instead.** User
-     *   specified, unique identifier.
      * @param {objectCallback} [callback] - Optional callback. Use instead of Promise return value as desired.
      * @returns {Promise<Key>} Newly created key.
      */
-    create: (params?: { id?: string; alias?: string }, cb?: ObjectCallback) => {
+    create: (params?: { id?: string }, cb?: ObjectCallback) => {
       return sharedAPI.tryCallback(client.request('/create-key', params), cb)
     },
 
@@ -51,8 +45,6 @@ export const keysAPI = (client: Client) => {
      *
      * @param {Object} params - Filter and pagination information.
      * @param {Array.<string>} params.ids - List of requested ids, max 200.
-     * @param {Array.<string>} params.aliases - **Deprecated. Use ids instead.**
-     *   List of requested aliases, max 200.
      * @param {Number} params.pageSize - Number of items to return in result set.
      * @param {PageCallback} [callback] - Optional callback. Use instead of Promise return value as desired.
      * @returns {Promise<Page<Key>>} Requested page of results.
@@ -60,12 +52,9 @@ export const keysAPI = (client: Client) => {
      * @deprecated Use {@link module:KeysApi~list|list} instead.
      */
     queryPage: (
-      params: { ids?: string[]; aliases?: string[]; pageSize?: number },
+      params: { ids?: string[]; pageSize?: number },
       cb?: PageCallback
     ) => {
-      if (Array.isArray(params.aliases) && params.aliases.length > 0) {
-        params.pageSize = params.aliases.length
-      }
       if (Array.isArray(params.ids) && params.ids.length > 0) {
         params.pageSize = params.ids.length
       }
@@ -86,8 +75,6 @@ export const keysAPI = (client: Client) => {
      *
      * @param {Object} params= - Filter and pagination information.
      * @param {Array.<string>} params.ids - List of requested ids, max 200.
-     * @param {Array.<string>} params.aliases - **Deprecated. Use ids instead.**
-     *   List of requested aliases, max 200.
      * @param {QueryProcessor<Account>} consumer - Processing callback.
      * @param {objectCallback} [callback] - Optional callback. Use instead of Promise return value as desired.
      * @returns {Promise} A promise resolved upon processing of all items, or
@@ -96,7 +83,7 @@ export const keysAPI = (client: Client) => {
      * @deprecated Use {@link module:KeysApi~list|list} instead.
      */
     queryEach: (
-      params: { ids?: string[]; aliases?: string[] },
+      params: { ids?: string[] },
       consumer: Consumer,
       cb?: ObjectCallback
     ) => sharedAPI.queryEach(client, 'keys', params, consumer, { cb }),
@@ -108,18 +95,14 @@ export const keysAPI = (client: Client) => {
      *
      * @param {Object} params - Filter and pagination information.
      * @param {Array.<string>} params.ids - List of requested ids, max 200.
-     * @param {Array.<string>} params.aliases - **Deprecated. Use ids instead.**
-     *   List of requested aliases, max 200.
      * @param {objectCallback} [callback] - Optional callback. Use instead of Promise return value as desired.
      * @returns {Promise} A promise resolved upon processing of all items, or
      *                   rejected on error.
      *
      * @deprecated Use {@link module:KeysApi~list|list} instead.
      */
-    queryAll: (
-      params?: { ids?: string[]; aliases?: string[] },
-      cb?: ObjectCallback
-    ) => sharedAPI.queryAll(client, 'keys', params, { cb }),
+    queryAll: (params?: { ids?: string[] }, cb?: ObjectCallback) =>
+      sharedAPI.queryAll(client, 'keys', params, { cb }),
 
     /**
      * Query a list of keys matching the specified query.